Create 1-ceaser-cipher.rb
Not ready for deployment. 
The idea is that every password is shifted in ascii/unicode (dependent on implementation, this is a prototype) value, by the length of the password. Then, the length of the password is added to the end. 

I.e. 

8 character password ABCDEFGH would be changed to 
BCDEFGHI8

This intended effect is to eradicate dictionary attacks against salted passwords.